hw3/pr3/requirements.txt
pandas==2.2.2
numpy==1.24.3
pyarrow==17.0.0
networkx==3.3
tables==3.10.1
h5py==3.9.0
netCDF4==1.6.3
xarray==2023.1.0
openpyxl==3.1.2
python-igraph==0.10.4